csv-cleaner/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
slowapi>=0.1.9
